+++ granulate_utils/config_feeder/client/client.py
@@ -1,21 +1,12 @@
 import asyncio
-import json
 import logging
-from collections import defaultdict
-from typing import DefaultDict, Dict, List, Optional, Union
+from typing import Callable, List, Optional, Union
 
-from granulate_utils.config_feeder.client.bigdata import get_node_info
-from granulate_utils.config_feeder.client.collector import ConfigFeederCollector
+from granulate_utils.config_feeder.client.collector import ConfigFeederCollector, ConfigFeederCollectorParams
 from granulate_utils.config_feeder.client.exceptions import ClientError
 from granulate_utils.config_feeder.client.http_client import HttpClient
-from granulate_utils.config_feeder.client.models import CollectionResult, ConfigType
-from granulate_utils.config_feeder.client.yarn.collector import YarnConfigCollector
-from granulate_utils.config_feeder.client.yarn.models import YarnConfig
-from granulate_utils.config_feeder.core.models.aggregation import CreateNodeConfigsRequest, CreateNodeConfigsResponse
-from granulate_utils.config_feeder.core.models.cluster import ClusterCreate, CreateClusterRequest, CreateClusterResponse
 from granulate_utils.config_feeder.core.models.collection import CollectorType
-from granulate_utils.config_feeder.core.models.node import CreateNodeRequest, CreateNodeResponse, NodeCreate, NodeInfo
-from granulate_utils.config_feeder.core.models.yarn import NodeYarnConfigCreate
+from granulate_utils.config_feeder.client.yarn_config_feeder_collector import YarnConfigFeederCollector
 
 
 class ConfigFeederClient:
@@ -28,119 +19,34 @@ class ConfigFeederClient:
         server_address: Optional[str] = None,
         yarn: bool = True,
         collector_type=CollectorType.SAGENT,
-        collectors: List[ConfigFeederCollector] = [],
+        collector_factories: List[Callable[[ConfigFeederCollectorParams], ConfigFeederCollector]] = [],
     ) -> None:
         if not token or not service:
             raise ClientError("Token and service must be provided")
+
+        def yarn_collector(params: ConfigFeederCollectorParams) -> YarnConfigFeederCollector:
+            return YarnConfigFeederCollector(params, yarn=yarn)
+
         self.logger = logger
         self._service = service
-        self._cluster_id: Optional[str] = None
         self._collector_type = collector_type
         self._is_yarn_enabled = yarn
-        self._yarn_collector = YarnConfigCollector(logger=logger)
-        self._last_hash: DefaultDict[ConfigType, Dict[str, str]] = defaultdict(dict)
-        self._collectors = collectors
         self._http_client = HttpClient(token, server_address)
+        self._collectors = self._create_collectors([yarn_collector, *collector_factories])
 
     def collect(self) -> None:
-        if (node_info := get_node_info(self.logger)) is None:
-            self.logger.warning("not a Big Data host, skipping")
-            return None
+        asyncio.run(self._collect())
 
-        collection_result = asyncio.run(self._collect(node_info))
+    async def _collect(self) -> None:
+        await asyncio.gather(*list(map(lambda c: c.collect(), self._collectors)))
 
-        if self._cluster_id is None and (node_info.is_master or not collection_result.is_empty):
-            self._register_cluster(node_info)
-
-        if collection_result.is_empty:
-            self.logger.info("no configs to submit")
-            return None
-
-        self._submit_node_configs(collection_result)
-
-    async def _collect(self, node_info: NodeInfo) -> CollectionResult:
-        results = await asyncio.gather(
-            self._collect_yarn_config(node_info),
-        )
-        return CollectionResult(node=node_info, yarn_config=results[0])
-
-    async def _collect_yarn_config(self, node_info: NodeInfo) -> Optional[YarnConfig]:
-        if not self._is_yarn_enabled:
-            return None
-        self.logger.info("YARN config collection starting")
-        yarn_config = await self._yarn_collector.collect(node_info)
-        self.logger.info("YARN config collection finished")
-        return yarn_config
-
-    def _submit_node_configs(
-        self,
-        collection_result: CollectionResult,
-    ) -> None:
-        external_id = collection_result.node.external_id
-        request = self._get_configs_request(collection_result)
-        if request is None:
-            self.logger.info(f"skipping node {external_id}, configs are up to date")
-            return None
-
-        node_id = self._register_node(collection_result.node)
-        self.logger.info(f"sending configs for node {external_id}")
-        response = CreateNodeConfigsResponse(**self._http_client.request("POST", f"/nodes/{node_id}/configs", request))
-
-        if response.yarn_config is not None:
-            assert request.yarn_config is not None
-            self._last_hash[ConfigType.YARN][external_id] = collection_result.yarn_config_hash
-
-    def _register_node(
-        self,
-        node: NodeInfo,
-    ) -> str:
-        assert self._cluster_id is not None
-        self.logger.debug(f"registering node {node.external_id}")
-        request = CreateNodeRequest(
-            node=NodeCreate(
-                collector_type=self._collector_type,
-                external_id=node.external_id,
-                is_master=node.is_master,
-            ),
-            allow_existing=True,
-        )
-        response = CreateNodeResponse(
-            **self._http_client.request("POST", f"/clusters/{self._cluster_id}/nodes", request)
-        )
-        return response.node.id
-
-    def _register_cluster(self, node_info: NodeInfo) -> None:
-        self.logger.debug(f"registering cluster {node_info.external_id}")
-        request = CreateClusterRequest(
-            cluster=ClusterCreate(
-                collector_type=self._collector_type,
-                service=self._service,
-                provider=node_info.provider,
-                bigdata_platform=node_info.bigdata_platform,
-                external_id=node_info.external_cluster_id,
-                properties=json.dumps(node_info.properties) if node_info.properties else None,
-            ),
-            allow_existing=True,
-        )
-        response = CreateClusterResponse.parse_obj(self._http_client.request("POST", "/clusters", request))
-        self._cluster_id = response.cluster.id
-
-    def _get_configs_request(self, configs: CollectionResult) -> Optional[CreateNodeConfigsRequest]:
-        yarn_config = self._get_yarn_config_if_changed(configs)
-
-        if yarn_config is None:
-            return None
-
-        return CreateNodeConfigsRequest(
-            yarn_config=yarn_config,
-        )
-
-    def _get_yarn_config_if_changed(self, configs: CollectionResult) -> Optional[NodeYarnConfigCreate]:
-        if configs.yarn_config is None:
-            return None
-        if self._last_hash[ConfigType.YARN].get(configs.node.external_id) == configs.yarn_config_hash:
-            self.logger.debug("YARN config is up to date")
-            return None
-        return NodeYarnConfigCreate(
-            collector_type=self._collector_type, config_json=json.dumps(configs.yarn_config.config)
-        )
+    def _create_collectors(
+        self, collector_factories: List[Callable[[ConfigFeederCollectorParams], ConfigFeederCollector]]
+    ):
+        params: ConfigFeederCollectorParams = {
+            "logger": self.logger,
+            "http_client": self._http_client,
+            "service": self._service,
+            "collector_type": self._collector_type,
+        }
+        return list(map(lambda fac: fac(params), collector_factories))
